+++ src/tokens/components/copyable-value.ts
@@ -1,0 +1,4 @@
+import copyableValue from '../../../tokens/component/copyable-value.json';
+import { flattenTokens } from '../utils';
+
+export const copyableValueTokens = flattenTokens(copyableValue as never);